Added more functions, refactored DatabaseService to allow as much unit testing as is feasible, added unit test for DatabaseService. Refactored ratings to be floats instead of integers, including rounding to 1 decimal place. Added acceptable bounds on range of possible ratings (0.0 - 5.0). Added edit movie functionality.

+++ src/main/java/com/mr/moviecatalogue/domain/Movie.java
@@ -14,5 +14,5 @@ import java.util.Optional;
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class Movie implements Serializable {
     private Optional<String> director;
-    private Optional<Integer> rating;
+    private Optional<Float> rating;
 }
